is() renamed to as()
